# backend-dotnet/api/openapi.yaml
openapi: 3.1.0
info:
  title: Claims API (life claim slice)
  version: 0.2.0
  summary: Individual life, disability income and annuity claims. This version covers one life claim from intake to closed (decision, payment run, close).
  description: |
    Derived from the React mock's `web/src/api` modules (their doc comments carry the verbs). Where the mock's shape is
    unsuitable for a real API the difference is deliberate; the full list is in `backend/README.md` ("Deviations from the mock").

    **Conventions**

    * **Ids** are UUIDs. Claims also have a human `claimNumber` (`L-26-043310`), unique, filterable, never used in paths.
    * **Money** is `{ "amount": "200000.00", "currency": "USD" }`: a decimal string with at most 2 decimals, never a JSON number.
    * **Timestamps** are RFC 3339 in UTC (`2026-09-25T15:03:00Z`). Business dates (date of death, issue date) are plain dates.
    * **Enum values** are snake_case, the same strings the database stores (and the deadline kinds in the architecture doc).
    * **Idempotency**: POSTs that create something take an `Idempotency-Key` header (8 to 128 characters). A repeat with the
      same key and body returns the resource it created (200, `Idempotent-Replayed: true`); the same key with a different body is 409.
    * **Optimistic concurrency**: every mutable resource has a `version`, returned as its `ETag` (`"3"`). State-changing calls on
      an existing resource require `If-Match`: missing is 428, stale is 412.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /requirements/{requirementId}:
    parameters: [{ $ref: '#/components/parameters/RequirementId' }]
    get:
      tags: [Requirements]
      operationId: getRequirement
      summary: Get a requirement
      description: One requirement, with its ETag for a later accept or waive.
      x-slice-status: implemented
      responses:
        '200':
          description: The requirement.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Requirement' }
        '404': { $ref: '#/components/responses/NotFound' }
  /requirements/{requirementId}:accept:
    parameters: [{ $ref: '#/components/parameters/RequirementId' }]
    post:
      tags: [Requirements]
      operationId: acceptRequirement
      summary: Accept a requirement
      description: |
        Postgres only, one transaction: the requirement becomes `accepted`, its waiting follow-up deadline rows close without
        firing, and if it was the last one, proof of loss is complete (status `in_review`, `decision_due` and `review_target`
        rows written, a work item for the owner). Mock: `markReceived` / `satisfyRequirement`. The event workflow that would call
        this when a document is classified is not built.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                satisfiedBy: { type: string, description: 'What satisfied it, shown on the requirement.' }
      responses:
        '200':
          description: The accepted requirement.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Requirement' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /requirements/{requirementId}:waive:
    parameters: [{ $ref: '#/components/parameters/RequirementId' }]
    post:
      tags: [Requirements]
      operationId: waiveRequirement
      summary: Waive a requirement, with a reason
      description: Mock `waiveRequirement`. Same effects as accept. A reason is required (422 `reason_required`).
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1 }
      responses:
        '200':
          description: The waived requirement.
          headers: { ETag: { $ref: '#/components/headers/ETag' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Requirement' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
...
  /claims/{claimId}/payment-items:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Payments]
      operationId: listClaimPaymentItems
      summary: Payment items for a claim
      description: Mock `GET /claims/{id}/payments`. Items are the contract between claims and payments. Oldest first.
      x-slice-status: implemented
      responses:
        '200':
          description: Items.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/PaymentItem' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
  /payment-items/{paymentItemId}:hold:
    parameters: [{ $ref: '#/components/parameters/PaymentItemId' }]
    post:
      tags: [Payments]
      operationId: holdPaymentItem
      summary: Hold a payment item
      description: |
        Only an item not yet in a run (`awaiting_proof` or `cleared`) can be held. A hold that arrives after the item is `in_run` is 409
        `item_in_run` and becomes a recovery work item instead; only the payment run touches an item in a run. A paid item is 409
        `invalid_state`. A held item is skipped by every run and keeps the claim open until it is released and paid.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason: { type: string, minLength: 1 }
      responses:
        '200':
          description: The held item.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentItem' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /payment-items/{paymentItemId}:release:
    parameters: [{ $ref: '#/components/parameters/PaymentItemId' }]
    post:
      tags: [Payments]
      operationId: releasePaymentItem
      summary: Release a held payment item
      description: A `held` item goes back to `cleared` and is paid by the next run whose date is on or after its `payOn`. Any other state is 409 `invalid_state`.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IfMatch'
        - $ref: '#/components/parameters/Actor'
      responses:
        '200':
          description: The released item.
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentItem' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '412': { $ref: '#/components/responses/PreconditionFailed' }
        '428': { $ref: '#/components/responses/PreconditionRequired' }
  /payment-runs:
    get:
      tags: [Payments]
      operationId: listPaymentRuns
      summary: Payment runs, newest first
      description: The daily runs and any manual runs.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Runs.
          content:
            application/json:
              schema:
                type: object
                required: [items, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/PaymentRun' } }
                  nextCursor: { type: [string, 'null'] }
        '400': { $ref: '#/components/responses/BadRequest' }
    post:
      tags: [Payments]
      operationId: startPaymentRun
      summary: Start a payment run by hand
      description: |
        The daily run has its own scheduler (an in-process trigger, `Claims:PaymentRun:*`) and does NOT use Temporal; this is the manual
        trigger. One request builds and finishes the run: it claims every `cleared` item with `payOn` on or before `runDate` in one statement
        (`cleared` to `in_run`, stamped with the run id, `FOR UPDATE SKIP LOCKED`, so two runs never take the same item), sends the file
        to the bank (a stub here), marks the items `paid` (or `returned`), and closes every claim whose items are all paid. Because it is
        synchronous the response is the finished run (`reconciled`, or `failed` when the bank could not be reached; the items then go back
        to `cleared`). `runDate` cannot be in the future of the business date (422 `run_date_in_future`).

        Idempotent: a repeat with the same `Idempotency-Key` returns the run (200, `Idempotent-Replayed: true`). Without the header the key
        defaults to the run date, so a second click for the same date returns the first run; send a new key to force another run.
      x-slice-status: implemented
      parameters:
        - { name: Idempotency-Key, in: header, required: false, description: 'Optional here; defaults to `manual-<runDate>`.', schema: { type: string, minLength: 8, maxLength: 128 } }
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [runDate]
              properties:
                runDate: { type: string, format: date }
      responses:
        '201':
          description: The run was created and has finished.
          headers:
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentRun' }
        '200':
          description: The same key was already used; the run it created.
          headers:
            Idempotent-Replayed: { schema: { type: string, const: 'true' } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentRun' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
  /payment-runs/{paymentRunId}:
    parameters: [{ $ref: '#/components/parameters/PaymentRunId' }]
    get:
      tags: [Payments]
      operationId: getPaymentRun
      summary: Get a payment run
      description: One run with its totals and status.
      x-slice-status: implemented
      responses:
        '200':
          description: The run.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentRun' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /payment-runs/returns:process:
    post:
      tags: [Payments]
      operationId: processBankReturns
      summary: The returns batch (also run by the daily job)
      description: |
        A BATCH, not a workflow. Reads the bank's returns queue (what the returns file would carry: item id, code, reason), and for each item that is
        `paid` marks it `returned` (the original stays as it was: paid items are immutable; a return only adds the code and reason), records a
        payment-method hold on that payee's closed account, appends the history, and writes a `payment_returned` outbox event, in one transaction.
        Items are locked with `FOR UPDATE SKIP LOCKED`, so two batches at once never process the same return, and an item already returned is skipped.
        The batch NEVER changes the claim: `PaymentReturnedWorkflow` (`event-<outbox event id>`) reopens the claim, letters the payee and tells the
        examiner. Returns are acknowledged to the bank only after the commit, so a crash re-reads them harmlessly.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/Actor'
      responses:
        '200':
          description: What the batch did.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ReturnsProcessed' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /claims/{claimId}/payment-methods:
    parameters: [{ $ref: '#/components/parameters/ClaimId' }]
    get:
      tags: [Payments]
      operationId: listClaimPaymentMethods
      summary: Payment methods a payee added on this claim, and the holds on closed accounts
      x-slice-status: implemented
      responses:
        '200':
          description: Methods (oldest first) and holds.
          content:
            application/json:
              schema:
                type: object
                required: [items, holds, nextCursor]
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/PaymentMethod' } }
                  holds: { type: array, items: { $ref: '#/components/schemas/PaymentMethodHold' } }
                  nextCursor: { type: [string, 'null'] }
        '404': { $ref: '#/components/responses/NotFound' }
  /claims/{claimId}/payees/{partyId}:update-payment-method:
    parameters:
      - $ref: '#/components/parameters/ClaimId'
      - $ref: '#/components/parameters/PartyId'
    post:
      tags: [Payments]
      operationId: updatePayeePaymentMethod
      summary: A payee whose payment came back gives a new account
      description: |
        ONE transaction (saved before any workflow): the new method is stored (only the last 4 digits of the routing and account numbers are kept),
        the payee's `bank_details_by` deadline row closes, the REPLACEMENT payment item is created (`replacementOfId` = the returned item, the same
        principal and interest: interest stopped at the first payment date; `cleared`, `payOn` the next business day, linked to the new method), a
        `payment_due` "reissue target" row (2 business days) is written, the examiner's "payment returned" work item closes, history is appended, and an
        outbox event `payment_method_updated` is written. The next payment run pays the replacement and, when it is the last unpaid item, closes the
        claim again. The short workflow `PaymentMethodUpdatedWorkflow` (`event-<outbox event id>`) verifies the account with the bank (stub) and confirms
        to the payee (RTN-LIFE-02); if the account cannot be verified it holds the replacement item and opens a work item.
        409 `no_returned_payment` when the payee has no returned payment without a replacement. Idempotent by `Idempotency-Key`.
      x-slice-status: implemented
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/Actor'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [kind, routingNumber, accountNumber]
              properties:
                kind: { type: string, enum: [eft] }
                routingNumber: { type: string, pattern: '^[0-9]{9}$' }
                accountNumber: { type: string, pattern: '^[0-9]{4,17}$' }
                holderName: { type: string }
      responses:
        '201':
          description: The stored method and the replacement item(s).
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentMethodUpdated' }
        '200':
          description: 'Same key and body as an earlier call (`Idempotent-Replayed: true`).'
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PaymentMethodUpdated' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '422': { $ref: '#/components/responses/Unprocessable' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'The intended scheme (org identity provider). NOT enforced yet: the service is unauthenticated and trusts X-Actor. See the README, not built yet.'
  headers:
    ETag:
      description: The resource's version, quoted. Send it back as `If-Match`.
      schema: { type: string, examples: ['"3"'] }
  parameters:
    ClaimId: { name: claimId, in: path, required: true, schema: { type: string, format: uuid } }
    RequirementId: { name: requirementId, in: path, required: true, schema: { type: string, format: uuid } }
...
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: A client-generated unique key (for example a UUID), 8 to 128 characters.
      schema: { type: string, minLength: 8, maxLength: 128 }
    IfMatch:
      name: If-Match
      in: header
      required: true
      description: The ETag from the last GET of this resource.
      schema: { type: string, examples: ['"3"'] }
...
    PaymentItemId: { name: paymentItemId, in: path, required: true, schema: { type: string, format: uuid } }
    PaymentRunId: { name: paymentRunId, in: path, required: true, schema: { type: string, format: uuid } }
...
    PartyId: { name: partyId, in: path, required: true, schema: { type: string, format: uuid } }
...
    Actor:
      name: X-Actor
      in: header
      required: false
      description: Development placeholder for the acting user until authentication is built.
      schema: { type: string }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
...
  responses:
    BadRequest:
      description: The request could not be read (malformed JSON, bad header or cursor).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
...
    NotFound:
      description: No such resource.
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Conflict:
      description: The request conflicts with the resource's state (`invalid_state`, `idempotency_key_reuse`, `deadline_not_open`, `claim_not_ready_to_decide`, `not_awaiting_approval`, `item_in_run`, ...).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    PreconditionFailed:
      description: '`If-Match` is stale (`version_conflict`). Fetch the resource again.'
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    PreconditionRequired:
      description: '`If-Match` is required (`precondition_required`).'
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
    Unprocessable:
      description: Well formed but not valid (`validation_failed` with field errors, or a business-rule code such as `payee_shares_invalid`).
      content: { application/problem+json: { schema: { $ref: '#/components/schemas/Problem' } } }
  schemas:
...
    RequirementState:
      type: string
      enum: [requested, received, accepted, not_enough, waived, expired]
...
    Requirement:
      type: object
      required: [id, claimId, key, name, purpose, from, state, requestedAt, followUpDays, followUpCount, version]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        benefitLineId: { type: [string, 'null'], format: uuid }
        key: { type: string, enum: [certificate, statement, primary_died_first, report, amended_certificate] }
        name: { type: string }
        purpose: { type: string, description: 'What it proves: "Proof of death", "Proof of claim".' }
        from:
          type: object
          required: [label]
          properties:
            partyId: { type: [string, 'null'], format: uuid }
            label: { type: string }
            detail: { type: [string, 'null'] }
        state: { $ref: '#/components/schemas/RequirementState' }
        stateNote: { type: [string, 'null'], description: 'Why it is in this state when that is not obvious: "TIN mismatch" for `not_enough`, "Under review: photocopy" for `received`.' }
        requestedAt: { type: string, format: date-time }
        followUpDays: { type: integer, description: Days until the follow-up deadline fires; 0 means never chased. }
        followUpCount: { type: integer }
        lastReminderAt: { type: [string, 'null'], format: date-time }
        receivedAt: { type: [string, 'null'], format: date-time }
        acceptedAt: { type: [string, 'null'], format: date-time }
        satisfiedBy: { type: [string, 'null'] }
        waivedAt: { type: [string, 'null'], format: date-time }
        waiveReason: { type: [string, 'null'] }
        version: { type: integer, format: int64 }
...
    PaymentItem:
      type: object
      required: [id, claimId, benefitLineId, kind, payeePartyId, basis, principal, interest, amount, method, payOn, status, version]
      properties:
        id: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        benefitLineId: { type: string, format: uuid }
        decisionId: { type: [string, 'null'], format: uuid }
        kind: { type: string, enum: [benefit, adjustment] }
        adjustsItemId: { type: [string, 'null'], format: uuid, description: 'For an adjustment, the item it corrects. Paid items are never edited.' }
        replacementOfId: { type: [string, 'null'], format: uuid, description: 'For a replacement of a returned item: the returned item it replaces (a new item; the returned one is never edited).' }
        replacedById: { type: [string, 'null'], format: uuid, description: 'For a returned item, the replacement item created for it.' }
        paymentMethodId: { type: [string, 'null'], format: uuid, description: 'The payee''s new account this item pays to; null means the account on file at intake.' }
        payeePartyId: { type: string, format: uuid }
        payeeName: { type: string }
        basis: { type: string, examples: ['50% of $200,000.00 + $191.78 interest'] }
        principal: { $ref: '#/components/schemas/Money' }
        interest: { $ref: '#/components/schemas/Money' }
        amount: { $ref: '#/components/schemas/Money' }
        method: { type: string, enum: [eft, check], description: 'EFT, or a check for a payee whose packet goes by mail.' }
        payOn: { type: string, format: date, description: 'The earliest run date that may pay it.' }
        status:
          type: string
          enum: [awaiting_proof, cleared, in_run, paid, held, cancelled, returned]
          description: 'awaiting_proof, cleared, in_run, paid; held (by a person), cancelled, returned (by the bank). Paid items never change; a return is followed by a new replacement item.'
        holdReason: { type: [string, 'null'] }
        runId: { type: [string, 'null'], format: uuid }
        paidAt: { type: [string, 'null'], format: date-time }
        paymentReference: { type: [string, 'null'], description: 'The bank''s reference for a paid item.' }
        returnCode: { type: [string, 'null'], description: 'The bank''s reason code when returned (R02 account closed).' }
        returnReason: { type: [string, 'null'] }
        version: { type: integer, format: int64 }
    PaymentRun:
      type: object
      required: [id, runDate, status, trigger, itemCount, total, startedAt]
      properties:
        id: { type: string, format: uuid }
        runDate: { type: string, format: date }
        status: { type: string, enum: [building, sent, reconciled, failed] }
        trigger: { type: string, enum: [schedule, manual] }
        itemCount: { type: integer }
        paidCount: { type: integer }
        returnedCount: { type: integer }
        total: { $ref: '#/components/schemas/Money' }
        fileReference: { type: [string, 'null'] }
        error: { type: [string, 'null'], description: 'Why the run failed (the bank was unreachable).' }
        startedAt: { type: string, format: date-time }
        finishedAt: { type: [string, 'null'], format: date-time }
...
    ReturnsProcessed:
      type: object
      required: [readAt, read, processed, skipped]
      properties:
        readAt: { type: string, format: date-time }
        read: { type: integer, description: Returns read from the bank's queue. }
        processed:
          type: array
          items:
            type: object
            required: [paymentItemId, claimId, payeeName, amount, returnCode, outboxEventId]
            properties:
              paymentItemId: { type: string, format: uuid }
              claimId: { type: string, format: uuid }
              payeeName: { type: string }
              amount: { $ref: '#/components/schemas/Money' }
              returnCode: { type: string }
              returnReason: { type: [string, 'null'] }
              outboxEventId: { type: string, format: uuid, description: 'The `payment_returned` event; its workflow is `event-<this id>`.' }
        skipped:
          type: array
          items:
            type: object
            required: [paymentItemId, reason]
            properties:
              paymentItemId: { type: string, format: uuid }
              reason: { type: string, enum: [not_found, not_paid, locked_by_another_batch], description: 'not_paid covers an item already returned (a repeat) or one that was never paid.' }
    PaymentMethod:
      type: object
      required: [id, partyId, claimId, kind, routingLast4, accountLast4, status, createdAt, version]
      properties:
        id: { type: string, format: uuid }
        partyId: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        kind: { type: string, enum: [eft] }
        routingLast4: { type: string }
        accountLast4: { type: string }
        holderName: { type: [string, 'null'] }
        status: { type: string, enum: [pending_verification, verified, rejected] }
        createdAt: { type: string, format: date-time }
        verifiedAt: { type: [string, 'null'], format: date-time }
        version: { type: integer, format: int64 }
    PaymentMethodHold:
      type: object
      required: [id, partyId, claimId, sourceItemId, reason, createdAt]
      description: 'The payee''s account on file (the one the payment went to) is closed: no item that pays to it is paid, while the hold lasts.'
      properties:
        id: { type: string, format: uuid }
        partyId: { type: string, format: uuid }
        claimId: { type: string, format: uuid }
        sourceItemId: { type: string, format: uuid, description: The returned payment item that caused the hold. }
        reason: { type: string, examples: ['R02 Account closed'] }
        createdAt: { type: string, format: date-time }
        supersededByMethodId: { type: [string, 'null'], format: uuid, description: The new account that replaced it. }
    PaymentMethodUpdated:
      type: object
      required: [paymentMethod, replacementItems]
      properties:
        paymentMethod: { $ref: '#/components/schemas/PaymentMethod' }
        replacementItems: { type: array, items: { $ref: '#/components/schemas/PaymentItem' } }
        reissueDeadline: { oneOf: [{ $ref: '#/components/schemas/Deadline' }, { type: 'null' }], description: 'The 2-business-day "reissue a returned payment" row.' }
...